+++ core/op/naive/permute.cpp
@@ -11,7 +11,7 @@ IMPL_NAIVE_SINGLE_INPUT_INTERNAL(permute) {
   size_t src_idx[NN_MAX_NDIM], dst_idx[NN_MAX_NDIM];
   for (size_t i = 0; i < n; i++) {
     loup.offset_to_indices(i, dst_idx);
-    for (size_t j = 0; j < loup.ndim; j++) {
+    for (int j = 0; j < loup.ndim; j++) {
       src_idx[j] = dst_idx[param.dims[j]];
     }
     size_t src_pos = linp.indices_to_offset(src_idx);
